refactor: extract widgets into reusable components
